requirements.txt
ccxt==3.1.54
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.0
ratelimit==2.2.1
schedule==1.2.0
matplotlib==3.7.2
scikit-learn==1.3.0
ta==0.10.2
requests==2.31.0
python-telegram-bot==20.5
backtrader==1.9.76.123
pytest==7.4.0
jsonschema==4.19.0
tqdm==4.66.1 